+++ src/arrays/allObjectives.ts
@@ -1,7 +1,9 @@
+import type { PlayerType } from "isaac-typescript-definitions";
 import { CHARACTER_OBJECTIVE_KINDS, OBJECTIVE_TYPES } from "../cachedEnums";
 import { DIFFICULTIES } from "../constants";
 import { ObjectiveType } from "../enums/ObjectiveType";
-import type { Objective } from "../types/Objective";
+import type { CharacterObjective, Objective } from "../types/Objective";
+import { getObjective } from "../types/Objective";
 import type { ObjectiveID } from "../types/ObjectiveID";
 import { getObjectiveID } from "../types/ObjectiveID";
 import { UNLOCKABLE_CHALLENGES } from "./unlockableChallenges";
@@ -14,17 +16,8 @@ export const ALL_OBJECTIVES: readonly Objective[] = (() => {
     switch (objectiveType) {
       case ObjectiveType.CHARACTER: {
         for (const character of PLAYABLE_CHARACTERS) {
-          for (const kind of CHARACTER_OBJECTIVE_KINDS) {
-            for (const difficulty of DIFFICULTIES) {
-              const objective: Objective = {
-                type: ObjectiveType.CHARACTER,
-                character,
-                kind,
-                difficulty,
-              };
-              objectives.push(objective);
-            }
-          }
+          const characterObjectives = getAllCharacterObjectives(character);
+          objectives.push(...characterObjectives);
         }
 
         break;
@@ -32,11 +25,11 @@ export const ALL_OBJECTIVES: readonly Objective[] = (() => {
 
       case ObjectiveType.CHALLENGE: {
         for (const challenge of UNLOCKABLE_CHALLENGES) {
-          const objective: Objective = {
-            type: ObjectiveType.CHALLENGE,
+          const challengeObjective = getObjective(
+            ObjectiveType.CHALLENGE,
             challenge,
-          };
-          objectives.push(objective);
+          );
+          objectives.push(challengeObjective);
         }
 
         break;
@@ -46,6 +39,26 @@ export const ALL_OBJECTIVES: readonly Objective[] = (() => {
 
   return objectives;
 })();
+
+export function getAllCharacterObjectives(
+  character: PlayerType,
+): CharacterObjective[] {
+  const characterObjectives: CharacterObjective[] = [];
+
+  for (const kind of CHARACTER_OBJECTIVE_KINDS) {
+    for (const difficulty of DIFFICULTIES) {
+      const characterObjective = getObjective(
+        ObjectiveType.CHARACTER,
+        character,
+        kind,
+        difficulty,
+      );
+      characterObjectives.push(characterObjective);
+    }
+  }
+
+  return characterObjectives;
+}
 
 export const ALL_OBJECTIVE_IDS: readonly ObjectiveID[] = ALL_OBJECTIVES.map(
   (objective) => getObjectiveID(objective),