+++ src/arrays/unlockablePickupTypes.ts
@@ -7,6 +7,7 @@ import {
   PickupVariant,
   SackSubType,
 } from "isaac-typescript-definitions";
+import { ReadonlySet } from "isaacscript-common";
 
 export const UNLOCKABLE_HEART_SUB_TYPES = [
   HeartSubType.GOLDEN, // 7
@@ -23,6 +24,10 @@ export const UNLOCKABLE_HEART_SUB_TYPES = [
   // - `HeartSubType.HALF` (2) is always unlocked.
 ] as const;
 
+export const UNLOCKABLE_HEART_SUB_TYPES_SET = new ReadonlySet<HeartSubType>(
+  UNLOCKABLE_HEART_SUB_TYPES,
+);
+
 export const UNLOCKABLE_COIN_SUB_TYPES = [
   CoinSubType.STICKY_NICKEL, // 6
   CoinSubType.DOUBLE_PACK, // 4
@@ -32,6 +37,10 @@ export const UNLOCKABLE_COIN_SUB_TYPES = [
   CoinSubType.GOLDEN, // 7
   // - `CoinSubType.PENNY` (1) is always unlocked.
 ] as const;
+
+export const UNLOCKABLE_COIN_SUB_TYPES_SET = new ReadonlySet<CoinSubType>(
+  UNLOCKABLE_COIN_SUB_TYPES,
+);
 
 export const UNLOCKABLE_BOMB_SUB_TYPES = [
   BombSubType.DOUBLE_PACK, // 2
@@ -43,12 +52,20 @@ export const UNLOCKABLE_BOMB_SUB_TYPES = [
   // - `BombSubType.GIGA` (7) is always unlocked.
 ] as const;
 
+export const UNLOCKABLE_BOMB_SUB_TYPES_SET = new ReadonlySet<BombSubType>(
+  UNLOCKABLE_BOMB_SUB_TYPES,
+);
+
 export const UNLOCKABLE_KEY_SUB_TYPES = [
   KeySubType.CHARGED, // 4
   KeySubType.DOUBLE_PACK, // 3
   KeySubType.GOLDEN, // 2
   // - `KeySubType.NORMAL` (1) is always unlocked.
 ] as const;
+
+export const UNLOCKABLE_KEY_SUB_TYPES_SET = new ReadonlySet<KeySubType>(
+  UNLOCKABLE_KEY_SUB_TYPES,
+);
 
 export const UNLOCKABLE_BATTERY_SUB_TYPES = [
   BatterySubType.MICRO, // 2
@@ -57,10 +74,18 @@ export const UNLOCKABLE_BATTERY_SUB_TYPES = [
   BatterySubType.GOLDEN, // 4
 ] as const;
 
+export const UNLOCKABLE_BATTERY_SUB_TYPES_SET = new ReadonlySet<BatterySubType>(
+  UNLOCKABLE_BATTERY_SUB_TYPES,
+);
+
 export const UNLOCKABLE_SACK_SUB_TYPES = [
   SackSubType.NORMAL, // 1
   SackSubType.BLACK, // 2
 ] as const;
+
+export const UNLOCKABLE_SACK_SUB_TYPES_SET = new ReadonlySet<SackSubType>(
+  UNLOCKABLE_SACK_SUB_TYPES,
+);
 
 export const UNLOCKABLE_CHEST_PICKUP_VARIANTS = [
   PickupVariant.SPIKED_CHEST, // 52
@@ -76,3 +101,6 @@ export const UNLOCKABLE_CHEST_PICKUP_VARIANTS = [
   // - `PickupVariant.OLD_CHEST` (55) is always unlocked (since it does not randomly spawn).
   // - `PickupVariant.MOMS_CHEST` (390) is always unlocked (since it does not randomly spawn).
 ] as const;
+
+export const UNLOCKABLE_CHEST_PICKUP_VARIANTS_SET =
+  new ReadonlySet<PickupVariant>(UNLOCKABLE_CHEST_PICKUP_VARIANTS);