+++ src/classes/features/achievementTracker/completedObjectives.ts
@@ -4,23 +4,26 @@ import type {
   PlayerType,
 } from "isaac-typescript-definitions";
 import { iRange } from "isaacscript-common";
-import { UNLOCKABLE_CHALLENGES } from "../../../arrays/unlockableChallenges";
+import { getAllCharacterObjectives } from "../../../arrays/allObjectives";
+import {
+  UNLOCKABLE_CHALLENGES,
+  UNLOCKABLE_CHALLENGES_SET,
+} from "../../../arrays/unlockableChallenges";
 import { PLAYABLE_CHARACTERS } from "../../../arrays/unlockableCharacters";
-import { CHARACTER_OBJECTIVE_KINDS } from "../../../cachedEnums";
 import type { CharacterObjectiveKind } from "../../../enums/CharacterObjectiveKind";
 import { ObjectiveType } from "../../../enums/ObjectiveType";
 import type { Objective } from "../../../types/Objective";
-import { getObjective, getObjectiveFromID } from "../../../types/Objective";
+import { getObjective } from "../../../types/Objective";
 import { getObjectiveID } from "../../../types/ObjectiveID";
 import { v } from "./v";
 
-// -------
-// Generic
-// -------
+// -----------------------------
+// Objective - Generic functions
+// -----------------------------
 
 export function isObjectiveCompleted(objective: Objective): boolean {
   const objectiveID = getObjectiveID(objective);
-  return v.persistent.completedObjectiveIDs.includes(objectiveID);
+  return v.persistent.completedObjectiveIDs.has(objectiveID);
 }
 
 // -------------------------------
@@ -30,20 +33,9 @@ export function isObjectiveCompleted(objective: Objective): boolean {
 export function isAllCharacterObjectivesCompleted(
   character: PlayerType,
 ): boolean {
-  const { completedObjectiveIDs } = v.persistent;
-
-  const completedObjectives = completedObjectiveIDs.map((objectiveID) =>
-    getObjectiveFromID(objectiveID),
-  );
-
-  const completedCharacterObjectives = completedObjectives.filter(
-    (objective) =>
-      objective.type === ObjectiveType.CHARACTER &&
-      objective.character === character,
-  );
-
-  return (
-    completedCharacterObjectives.length === CHARACTER_OBJECTIVE_KINDS.length
+  const characterObjectives = getAllCharacterObjectives(character);
+  return characterObjectives.every((characterObjective) =>
+    isObjectiveCompleted(characterObjective),
   );
 }
 
@@ -64,11 +56,7 @@ export function isCharacterObjectiveCompleted(
     kind,
     difficulty,
   );
-  const objectiveID = getObjectiveID(objective);
-
-  const { completedObjectiveIDs } = v.persistent;
-
-  return completedObjectiveIDs.includes(objectiveID);
+  return isObjectiveCompleted(objective);
 }
 
 // -------------------------------
@@ -76,17 +64,9 @@ export function isCharacterObjectiveCompleted(
 // -------------------------------
 
 export function isAllChallengeObjectivesCompleted(): boolean {
-  const { completedObjectiveIDs } = v.persistent;
-
-  const completedObjectives = completedObjectiveIDs.map((objectiveID) =>
-    getObjectiveFromID(objectiveID),
+  return UNLOCKABLE_CHALLENGES.every((challenge) =>
+    isChallengeObjectiveCompleted(challenge),
   );
-
-  const completedChallengeObjectives = completedObjectives.filter(
-    (objective) => objective.type === ObjectiveType.CHALLENGE,
-  );
-
-  return completedChallengeObjectives.length === UNLOCKABLE_CHALLENGES.length;
 }
 
 export function isChallengeRangeObjectivesCompleted(
@@ -100,14 +80,10 @@ export function isChallengeRangeObjectivesCompleted(
 }
 
 export function isChallengeObjectiveCompleted(challenge: Challenge): boolean {
-  if (!UNLOCKABLE_CHALLENGES.includes(challenge)) {
+  if (!UNLOCKABLE_CHALLENGES_SET.has(challenge)) {
     return true;
   }
 
   const objective = getObjective(ObjectiveType.CHALLENGE, challenge);
-  const objectiveID = getObjectiveID(objective);
-
-  const { completedObjectiveIDs } = v.persistent;
-
-  return completedObjectiveIDs.includes(objectiveID);
+  return isObjectiveCompleted(objective);
 }