+++ src/types/Achievement.ts
@@ -1,0 +1,7 @@
+import type { ObjectiveID } from "./ObjectiveID";
+import type { UnlockID } from "./UnlockID";
+
+export interface Achievement {
+  objectiveID: ObjectiveID;
+  unlockID: UnlockID;
+}